Fixes from review:
* Moved new elements and tests to the component.go and its test file.
* Changed the test which used the attribute package to stop it from potentially failing because of a failure in the other package.
